requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
pywebview>=4.0.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.0.0
python-docx>=0.8.11
PyMuPDF>=1.20.0
ultralytics>=8.0.0
torch>=2.0.0
opencv-python>=4.5.0
numpy>=1.21.0
requests>=2.25.0
